tyndyn/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
